Optimization for Bigger servers
It was checking items every 0.5 seconds to fill vmenu -- Now it checks if they are in store every 2.5 seconds if they are in store loads items for menu if not it checks again in 2.5 seconds. 
If in store triggers instore = true , when you leave store it triggers instore = false then just keeps checking if your in store to load list when you go back into the store.